backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose[cryptography]
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose[cryptography]